player movement and camera
gravity is up to -40.
player movement and jump have better numbers, should be jumping 4 tiles high and approx 8 block long now.
camera is sized to be 27x48 tiles -- THIS MIGHT NOT WORK CORRECTLY, have to change camera aspect ratio to properCam in Game tab. message daniel for details. working on finding a solution to that.